requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
umap-learn
